DAMIEN-349 Cache utility endpoints for admins
