+++ damien/routes.py
@@ -41,6 +41,7 @@ def register_routes(app):
 
     # Register API routes.
     import damien.api.auth_controller
+    import damien.api.cache_controller
     import damien.api.config_controller
     import damien.api.department_controller
     import damien.api.department_form_controller
